extract google loading & sign in/out to own file
`google.ts` can more-or-less be dropped into any project, except for the
how-to-store local data logic
